trigger the login on loginSuccess instead of relying timestamp
